photo upload upgrade performance

+++ app/src/main/java/com/reals/app/data/repository/PreparedProfilePhotoUploadLifecycle.kt
@@ -1,6 +1,7 @@
 package com.reals.app.data.repository
 
 import com.reals.app.core.media.PreparedProfilePhotoUpload
+import com.reals.app.core.media.PreparedUploadFileOwnership
 
 internal suspend fun <T> PreparedProfilePhotoUpload.useDeletingFile(
     block: suspend (PreparedProfilePhotoUpload) -> T,
@@ -8,5 +9,7 @@ internal suspend fun <T> PreparedProfilePhotoUpload.useDeletingFile(
     try {
         block(this)
     } finally {
-        file.delete()
+        if (fileOwnership == PreparedUploadFileOwnership.RepositoryOwned) {
+            file.delete()
+        }
     }
